H̴̳̞̹̦̜̗͖̼̮̮ͫ̆͑͊ͪ͂ͮͧ͂̐̊̍̀̇̉̑͘E̶̢̫͚͙̘͇͇̼̭̎̔͊ͯ̀ͅL̸̡̡̞̬̱͚̟̹̒̿̆̑̈́͂̂̂ͪ̆͗ͩ̊ͣ̌ͩͮ̂͞ͅL̴̨̬͔̘̬͚̤̗̭͎̟̮͓ͤ̿ͣ͗ͤͪ͑͢͢͞Ớ̴͓̥̹̻͔̺͒͑̒ͭͯͤ̅̒͛͊̈͛ͬ͐̊͛̂ͪ͠ ̛͇͈͍̘̙͈̣̞͚̘̻̞͓̮͙̺̻̮̍ͭ̄ͫ̓͂̂̋̑̃ͥ̓̔͌ͭ̄̉̚͝ͅH̲̝͈̙ͩ̂ͭͣ̃̀̍̽̾ͫͨ̊̕͡Ê̡̗̦̺̺͓̖̦̦̺͍͗̈́͌͗ͮͮ͛̾̋ͫ̄ͣ͜K̷̨̧͔͍̲͚̠̼̫̜̍̂̋̄̐͌ͪĄ̭͓̟̗͖͕̜͖̘͉̔ͨ͋ͮͦ̐̓͗̑͊̀͝F̄̽̄̉ͣ̉ͣͫͤ̌̇͗͋̿ͬ͏͍̖̪̯̥̭͔̙̼͇̥̗̟O͌̀̋̿͞҉͞͏̰̰͕̭̻̹͙̖̦̹̼̯F̛̟̪͕̩̦̩̞̼̰͉̰̩̥͇ͩͣ̐̋̓̇͊ͭͤ̀͘͞ added edit

+++ src/AppBundle/Controller/DefaultController.php
@@ -10,7 +10,6 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Extension\Core\Type\ButtonType;
 
 class DefaultController extends Controller
 {
@@ -20,15 +19,12 @@ class DefaultController extends Controller
     public function indexAction(Request $request)
     {
 
-        $blogRepository = $this->getDoctrine()
-            ->getRepository( 'AppBundle:BlogPost' );
-        $blogPosts = $blogRepository->findAll();
-
-        //$deleteButton->add( 'deletePost', ButtonType::class );
+        $blogPosts = $this->getDoctrine()
+            ->getRepository( 'AppBundle:BlogPost' )
+            ->findAll();
 
         return $this->render('default/index.html.twig', array(
             'blogPosts' => $blogPosts,
-            'deleteButton' => 'I am delete'
         ));
     }
 
@@ -70,9 +66,13 @@ class DefaultController extends Controller
         $deletionId = $request->query->get('postId');
 
         $em = $this->getDoctrine()->getManager();
-        $deletionObject = $em->getRepository('AppBundle:BlogPost')->find($deletionId);
-        $em->remove($deletionObject);
-        $em->flush();
+        $deletionObject = $em->getRepository('AppBundle:BlogPost')
+            ->find($deletionId);
+        if( $deletionObject ){
+            $em->remove($deletionObject);
+            $em->flush();
+        }
+        
 
 
 
@@ -84,10 +84,12 @@ class DefaultController extends Controller
      * @Route("/edit-blog-post", name="edit_blog_post")
      */
     public function editBlogPostAction( Request $request ){
+        $newPost = new BlogPost();
         $editId = $request->query->get('postId');
 
         $em = $this->getDoctrine()->getManager();
-        $editObject = $em->getRepository('AppBundle:BlogPost')->find($editId);
+        $editObject = $em->getRepository('AppBundle:BlogPost')
+            ->find($editId);
         
         $defaults = array(
             'title' => $editObject->getTitle(),
@@ -95,23 +97,24 @@ class DefaultController extends Controller
             'message' => $editObject->getMessage(),
             );
 
-        $newPost = new BlogPost();
         
-        $form = $this->createFormBuilder($newPost, $defaults)
-            ->add( 'title', TextType::class )
-            ->add( 'user', TextType::class )
-            ->add( 'message', TextareaType::class )
+        $form = $this->createFormBuilder( $newPost )
+            ->add( 'title', TextType::class, array( 'data' => $editObject->getTitle() ) )
+            ->add( 'user', TextType::class, array( 'data' => $editObject->getUser() ) )
+            ->add( 'message', TextareaType::class, array( 'data' => $editObject->getMessage() ) )
             ->add( 'save', SubmitType::class, array( 'label' => 'Compelete edit!' ) )
             ->getForm();
 
         $form->handleRequest($request);
 
         if( $form->isSubmitted()  && $form->isValid() ){
-            $newPost = $form->getData();
-            $newPost->setDate( new \DateTime('now') );
+            //$editObject = $form->getData();
+            $editObject->setMessage( $form->getData()->getMessage() );
+            $editObject->setUser( $form->getData()->getUser() );
+            $editObject->setTitle( $form->getData()->getTitle() );
+            $editObject->setDate( new \DateTime('now') );
 
             $em = $this->getDoctrine()->getManager();
-            $em->persist($newPost);
             $em->flush();
 
             return $this->redirectToRoute('homepage');
@@ -119,7 +122,8 @@ class DefaultController extends Controller
 
 
 
-        return $this->render('default/delete-blog-post.html.twig' );
-
+        return $this->render('default/edit-blog-post.html.twig', array(
+            'form' => $form->createView(),
+        ));
     }
 }
